requirements.txt
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
crcmod==1.7
cryptography==43.0.1
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
html2text==2024.2.26
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
MarkupSafe==2.1.5
numpy==1.26.4
oss2==2.18.6
outcome==1.3.0.post0
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
html2text~=2024.2.26
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Mako==1.3.5
alembic==1.13.2
python-dotenv==1.0.1
pymupdf==1.24.9
easygui==0.98.3
feedparser~=6.0.2
google~=3.0.0
pdfminer_six==20221105
pdftitle~=0.3
pypdf==4.3.1
pypdf2==2.0.0
pyperclip==1.9.0